More database/storage changes
Changed so that Data.get_users() does the unpickling and returns User objects directly.  We can pretty much use this format for the rest of our object types.

Still working on how users will get the current database when they log in.  Right now I'm just trying to have the whole db file sent for the initial log on, and then new entries will be parsed by the Receive() thread and then updated.