Made the entire application exit when the UI exits by changing the server and listener threads to daemons.
Whenever the program is loaded, the database is now downloaded from the client that is also hosting the server.  This is going to be important for when we're not just using localhost to connect.  Users can join at different times, and need a current copy of the data. It can be tested locally by running a second instance from a different directory without the database file 'db.db'.

Changed the Log In and Sign Up pages to use the Data class.  The pages now check and update the local database and broadcasts a new user to other clients. Currently keeping track of the current user with the attribute current_user of the Ui_MainWindow class.

Added a password attribute to the User class and implemented password checking on the login page.

Username of the currently signed in user displays on thr User Settings tab.

Made the update and delete methods static so that an instance of the Data class doesn't have to be created by the UI in order to call them.  We will need to make a change to the Receive class in togather_client.py to be able to broadcast updates to other clients.