Data.update_user() now broadcasts to other clients.
Changed name of variable for the 5th byte in message header from 'is_db' to 'msg_type'.  The 5th byte in the header specifies the type of message so that the server and the receiver thread in the client can handle the message appropriately.

The message type is specified by an optional parameter in Client.Send()

Current message types:
0: Default type when parameter is not specified in Client.Send(). The Receive class in the Client thread attempts to add the object to the database.
1: Used when sending the database to the server.
2: Used when a client requests a current copy of the database from the server.  The server sends a request for the database to the client it was started with, then sends that database to the requester.
3: Used when clients send an exit command to the server. Closes the connection.
4: The Receive class in the Client thread attempts to update the object to the database instead of adding it.
5: Not yet implemented.  Will be used to delete objects.