Data.delete_user() should now update other clients.
The Data.update_user() update from my last commit didn't work as intended, so I also made some more changes to that.

Also added overrides for equality and inequality operators to the User class, so that we can compare them easier.  We should probably do that for all the classes.